{{cookiecutter.repo_name}}/requirements-test.txt
Flask-Testing==0.8.1
testcontainers==3.7.0
